# Reduced memory allocations in readIndexRange()